feat(firmware_integrity): Add CRC 32
