+++ bootloader/CMakeLists.txt
@@ -57,7 +57,7 @@ target_sources(${BOOT_TARGET}
         ${CMAKE_CURRENT_SOURCE_DIR}/../common/src/core/system.c
         ${CMAKE_CURRENT_SOURCE_DIR}/../common/src/core/system_stm32f4xx.c
         ${CMAKE_CURRENT_SOURCE_DIR}/../common/src/core/uart.c
-        ${CMAKE_CURRENT_SOURCE_DIR}/../common/src/utils/crc8.c
+        ${CMAKE_CURRENT_SOURCE_DIR}/../common/src/utils/crc.c
 
 )
