+++ common/include/utils/crc.h
@@ -5,4 +5,6 @@
 
 uint8_t calculate_crc8(uint8_t* data, uint32_t length);
 
+uint32_t crc32(const uint8_t* data, uint32_t length);
+
 #endif
