--- common/src/utils/crc8.c
@@ -1,19 +1,0 @@
-#include "crc8.h"
-
-uint8_t calculate_crc8(uint8_t* data, uint32_t length)
-{
-    uint8_t crc = 0;
-
-    for (uint32_t i = 0; i < length; i++) {
-        crc ^= data[i];
-        for (uint8_t j = 0; j < 8; j++) {
-            if (crc & 0x80) {
-                crc = (crc << 1) ^ 0x07;
-            } else {
-                crc <<= 1;
-            }
-        }
-    }
-
-    return crc;
-}
